Backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.13
faiss-cpu>=1.9.0
python-multipart>=0.0.6
pandas>=2.1.4
openai>=1.10.0
python-dotenv>=1.0.0
pydantic>=2.5.3
tiktoken>=0.5.0